requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
colorama==0.4.4
coverage==5.3
distlib==0.3.1
Django==3.1.4
django-cors-headers==3.5.0
django-countries==7.0
django-mysql==3.9.0
djangorestframework==3.12.2
filelock==3.0.12
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
mysqlclient @ file:///C:/Users/Rasha-PC/Desktop/Finally/Thesis-Project/backend/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
pep8==1.7.1
Pillow==8.0.1
pipenv==2020.11.15
pycodestyle==2.6.0
pylint==2.6.0
PyMySQL==0.10.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.2.1
virtualenv-clone==0.5.4
wrapt==1.12.1
